feat(jit): add an extra level of indirection for long lived heap resources
For each data type, we create a pinned continuous heap buffer that
contains slots of raw pointer to the heap resources. By `pinned`, we
ensure that the slot address won't change during the entire lifetime of
JIT compiler.

This extra level of indirection allows us to "swap" heap pointer with
external pointer when necessary to reduce heap allocation and data copy

Signed-off-by: Zihan <zihanli0822@gmail.com>